+++ hmac-sha1.c
@@ -31,21 +31,21 @@
 
 void
 hmac_sha1_set_key(struct hmac_sha1_ctx *ctx,
-		  unsigned key_length, const uint8_t *key)
+		  size_t key_length, const uint8_t *key)
 {
   HMAC_SET_KEY(ctx, &nettle_sha1, key_length, key);
 }
 
 void
 hmac_sha1_update(struct hmac_sha1_ctx *ctx,
-		 unsigned length, const uint8_t *data)
+		 size_t length, const uint8_t *data)
 {
   sha1_update(&ctx->state, length, data);
 }
 
 void
 hmac_sha1_digest(struct hmac_sha1_ctx *ctx,
-		 unsigned length, uint8_t *digest)
+		 size_t length, uint8_t *digest)
 {
   HMAC_DIGEST(ctx, &nettle_sha1, length, digest);
 }
