+++ umac64.c
@@ -49,7 +49,7 @@ umac64_set_key (struct umac64_ctx *ctx, const uint8_t *key)
 
 void
 umac64_set_nonce (struct umac64_ctx *ctx,
-		  unsigned nonce_length, const uint8_t *nonce)
+		  size_t nonce_length, const uint8_t *nonce)
 {
   assert (nonce_length > 0);
   assert (nonce_length <= AES_BLOCK_SIZE);
@@ -72,7 +72,7 @@ umac64_set_nonce (struct umac64_ctx *ctx,
 
 void
 umac64_update (struct umac64_ctx *ctx,
-	       unsigned length, const uint8_t *data)
+	       size_t length, const uint8_t *data)
 {
   MD_UPDATE (ctx, length, data, UMAC64_BLOCK, (void)0);
 }
@@ -80,7 +80,7 @@ umac64_update (struct umac64_ctx *ctx,
 
 void
 umac64_digest (struct umac64_ctx *ctx,
-	       unsigned length, uint8_t *digest)
+	       size_t length, uint8_t *digest)
 {
   uint32_t tag[2];
   uint32_t *pad;
